refactor a8w4smooth tests into self-contained module and fix const_expr guards
Move all a8w4smooth test helpers and runners from test_moe_gemm.py into
test_a8w4_moe_gemm_2stage.py as a standalone module. Revert test_moe_gemm.py
to main. Rename shadowed make_preshuffle_b_layout to
make_preshuffle_b_layout_packed4bit and add 90 missing const_expr() wrappers
in a8w4_moe_gemm_2stage.py kernel functions to prevent scf.if miscompilation.
